% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{as.data.frame.pandas.core.frame.DataFrame}
\alias{as.data.frame.pandas.core.frame.DataFrame}
\title{Convert from Pandas DataFrame}
\usage{
\method{as.data.frame}{pandas.core.frame.DataFrame}(x, convert_unicode = TRUE,
  ...)
}
\arguments{
\item{convert_unicode}{whether to convert string columns to unicode so R can read back correctly}
}
\description{
Convert from Pandas DataFrame
}
